requirements.txt
requests
beautifulsoup4
pymysql
yagmail
python-dotenv
